requirements.txt
torch>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
pygame>=2.5.0
PyOpenGL>=3.1.6
pytest>=7.0.0
